Adaptação de equivalencias
