+++ dicionario/Model/ConectaBanco.cs
@@ -87,8 +87,8 @@ namespace dicionario.Model
         }
     }
     class CRUD{
-        private ConectaBanco ControllerBanco = new ConectaBanco("lexdbase","lexdbase","Int3rl3x1c0gr@", "lexdbase.mysql.dbaas.com.br");
-        //private ConectaBanco ControllerBanco = new ConectaBanco();
+        //private ConectaBanco ControllerBanco = new ConectaBanco("lexdbase","lexdbase","Int3rl3x1c0gr@", "lexdbase.mysql.dbaas.com.br");
+        private ConectaBanco ControllerBanco = new ConectaBanco();
         private void EnviaComando(string query){
             if (ControllerBanco.AbreConexao() == true)
             {
